update list_bill_detail: onclick="return confirm()

+++ admin/list_bill_detail.php
@@ -56,7 +56,7 @@ $db = db::getInstance();
                         <td><?php echo $obj->price; ?></td>
                         <td><a href="edit_bill_detail.php?id=<?php echo $obj->id ?>"><img width="16px" src="../public/images/icon_edit.png"></a>
                         </td>
-                        <td><a href="delete_bill_detail.php?id=<?php echo $obj->id ?>"><img width="16px" src="../public/images/icon_delete.png"></a>
+                        <td><a href="delete_bill_detail.php?id=<?php echo $obj->id ?>" onclick="return confirm('Bạn có thực sự muốn xóa không?');"><img width="16px" src="../public/images/icon_delete.png"></a>
                         </td>
                     </tr>
                     <?php
